bugfix/FT-01: structurized handlers
