+++ src/js/handlers/onHomeLinkClick.js
@@ -1,0 +1,9 @@
+import { renderHomeHeader } from '../templates/header/renderHomeHeader';
+
+const onHomeLinkClick = event => {
+  event.preventDefault();
+
+  renderHomeHeader();
+};
+
+export { onHomeLinkClick };
