+++ src/js/handlers/onLibraryLinkClick.js
@@ -1,0 +1,9 @@
+import { renderLibraryHeader } from '../templates/header/renderLibraryHeader';
+
+const onLibraryLinkClick = event => {
+  event.preventDefault();
+
+  renderLibraryHeader();
+};
+
+export { onLibraryLinkClick };
