+++ src/js/handlers/onSearchMovie.js
@@ -1,0 +1,22 @@
+import Api from '../api-service';
+
+const appService = new Api();
+
+const onSearchMovie = e => {
+  e.preventDefault();
+
+  appService.query = e.currentTarget.elements.searchQuery.value.trim();
+
+  const isEmptySearch = appService.searchQuery === '';
+
+  if (isEmptySearch) {
+    console.log('Пустий пошук');
+    //Зробити повідомлення про пустий пошук
+    return;
+  }
+
+  appService.resetPage();
+  appService.fetchSearchMovie().then(console.log);
+};
+
+export { onSearchMovie, appService };
